+++ app/src/main/java/de/baumann/pdfcreator/pages/create_image.java
@@ -260,8 +260,6 @@ public class create_image extends Fragment {
         if (Intent.ACTION_SEND.equals(action) && type != null) {
             if (type.startsWith("image/")) {
                 handleSendImage(intent); // Handle single image being sent
-            } else if (type.startsWith("application/pdf")) {
-                handleSendPDF(intent); // Handle single image being sent
             }
         }
 
@@ -292,16 +290,6 @@ public class create_image extends Fragment {
                 e.printStackTrace();
             }
         }
-    }
-
-    private void handleSendPDF(Intent intent) {
-        Uri pdfUri = intent.getParcelableExtra(Intent.EXTRA_STREAM);
-
-        String FilePath = pdfUri.getPath();
-        String FileTitle = FilePath.substring(FilePath.lastIndexOf("/")+1);
-        sharedPref.edit().putString("pathPDF", FilePath).apply();
-        sharedPref.edit().putString("title", FileTitle).apply();
-        helper_pdf.pdf_textField(getActivity(), rootView);
     }
 
     private void createPDF() {
@@ -552,11 +540,29 @@ public class create_image extends Fragment {
     public void onResume() {
         super.onResume();    //To change body of overridden methods use File | Settings | File Templates.
         File imgFile = new File(Environment.getExternalStorageDirectory() + "/Pictures/.pdf_temp/pdf_temp.jpg");
+        helper_pdf.pdf_textField(getActivity(), rootView);
+        helper_pdf.toolbar(getActivity());
         if(imgFile.exists()){
             Bitmap myBitmap = BitmapFactory.decodeFile(imgFile.getAbsolutePath());
             img.setImageBitmap(myBitmap);
         } else {
             img.setImageResource(R.drawable.image);
+        }
+    }
+
+    @Override
+    public void setUserVisibleHint(boolean isVisibleToUser) {
+        super.setUserVisibleHint(isVisibleToUser);
+        if (isVisibleToUser && isResumed()) {
+            File imgFile = new File(Environment.getExternalStorageDirectory() + "/Pictures/.pdf_temp/pdf_temp.jpg");
+            helper_pdf.pdf_textField(getActivity(), rootView);
+            helper_pdf.toolbar(getActivity());
+            if(imgFile.exists()){
+                Bitmap myBitmap = BitmapFactory.decodeFile(imgFile.getAbsolutePath());
+                img.setImageBitmap(myBitmap);
+            } else {
+                img.setImageResource(R.drawable.image);
+            }
         }
     }